Move `SecureStore` to production
